requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-redis==5.4.0
django-environ==0.11.2
django-filter==24.3

psycopg2-binary==2.9.9
django-debug-toolbar==4.3.0


celery==5.4.0
redis==5.0.4

yfinance==0.2.40
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1

drf-yasg==1.21.7
reportlab==4.1.0
django-silk==5.0.4
gunicorn==20.1.0
